fastapi/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.115.0
h11==0.14.0
idna==3.10
pillow==10.4.0
pydantic==2.9.2
pydantic_core==2.23.4
python-multipart==0.0.10
requests==2.32.3
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
